last modification on code with best score on tiny/small

+++ instance.h
@@ -27,7 +27,7 @@ class Instance {
 
   float c_Pb, c_Ab, c_Db, c_Pp, c_Ap, c_Dp, c_1r, c_2r, u_P, u_A, cu;
 
-  vector<pair<int, pair<float, float>>> clients;
+  vector<pair<float, pair<float, float>>> clients;
   vector<pair<float, float>> sites;
   vector<vector<float>> siteSiteDistances;
   vector<vector<float>> siteClientDistances;
